+++ README.md
@@ -1,1 +1,7 @@
 Tasks-Compsite
+
+Паттерн Compsite
+
+На первом экране можно добавить только задачи
+На остальных экранах можно добавить задачи и подзадачи
+В подзадачу ничего нельзя добавить
