+++ Tasks-Compsite/Model/Component.swift
@@ -50,44 +50,3 @@ extension Component {
     return "\(parent.path())\\\(parent.name)"
   }
 }
-
-class SubTask: Component {
-  var name: String
-  var parent: Component?
-  
-  init(name: String) {
-    self.name = name
-  }
-}
-
-class Task: Component {
-  var name: String
-  var parent: Component?
-  var children = [Component]()
-  
-  init(name: String) {
-    self.name = name
-  }
-
-  func add(component: Component) {
-    var item = component
-    item.parent = self
-    children.append(item)
-  }
-  
-  func isTask() -> Bool {
-    return true
-  }
-  
-  func countTasks() -> Int {
-    children.reduce(0) { $0 + $1.countTasks() }
-  }
-  
-  func countElements() -> Int {
-    return children.count
-  }
-  
-  func getSubTask(for index: Int) -> Component? {
-    children[index]
-  }
-}
